% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_list.R
\name{as.data.frame.sim_list}
\alias{as.data.frame.sim_list}
\title{Convert a collection of spatial interaction model into a Data Frame}
\usage{
\method{as.data.frame}{sim_list}(x, ..., models = TRUE)
}
\arguments{
\item{x}{a collection of spatial interaction models, an object of class
\code{sim_list}}

\item{...}{additional parameters (not used currently)}

\item{models}{whether to include the models in the data frame (defaults to
TRUE, that is to model inclusion)}
}
\value{
a data frame representation of the spatial interaction model collection
}
\description{
This function build a data frame from a collection of spatial interaction model.
}
\details{
The data frame has one row per spatial interaction model and the following columns:
\itemize{
\item \code{alpha}: the return to scale parameter used to build the model
\item \code{beta}: the cost inverse scale parameter used to build the model
\item \code{diversity}: model diversity
\item \code{terminals}: if the origin and destination locations are identical, the
terminals of the model
\item \code{model}: if \code{models=TRUE}, the model
}
}
\examples{
positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
all_flows <- grid_blvim(distances, production, c(1.25, 1.5), c(1, 2, 3), attractiveness)
as.data.frame(all_flows, models = FALSE)

}
