% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{french_departments}
\alias{french_departments}
\title{French departments}
\format{
A \code{data.frame} with 101 rows and 3 columns
\describe{
\item{department}{The code of the department in the official geographical
code of 2025 (OGC)}
\item{region}{The code of the region of the city in the OGC (see \link{french_regions})}
\item{name}{The name of the department}
}
}
\source{
INSEE Official Geographical Code (2025)
\url{https://www.data.gouv.fr/datasets/code-officiel-geographique-cog/}
\url{https://www.data.gouv.fr/api/1/datasets/r/91a95bee-c7c8-45f9-a8aa-f14cc4697545}
}
\usage{
french_departments
}
\description{
This data set contains the list of all French departments. It can be joined
with the \link{french_regions} or the \link{french_cities} data set. The data set was
extracted from the INSEE open data in November 2025. These data are
distributed under the
\href{https://www.etalab.gouv.fr/wp-content/uploads/2014/05/Open_Licence.pdf}{French
"Open Licence"}.
}
\seealso{
\link{french_cities} and \link{french_regions}
}
\keyword{datasets}
