% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{french_regions}
\alias{french_regions}
\title{French regions}
\format{
A \code{data.frame} with 18 rows and 2 columns
\describe{
\item{region}{The code of the region in the official geographical
code of 2025 (OGC)}
\item{name}{The name of the region}
}
}
\source{
INSEE Official Geographical Code (2025)
\url{https://www.data.gouv.fr/datasets/code-officiel-geographique-cog/}
\url{https://www.data.gouv.fr/api/1/datasets/r/91a95bee-c7c8-45f9-a8aa-f14cc4697545}
}
\usage{
french_regions
}
\description{
This data set contains the list of all French regions It can be joined with
the \link{french_departments} or the \link{french_cities} data set. The data set was
extracted from the INSEE open data in November 2025. These data are
distributed under the \href{https://www.etalab.gouv.fr/wp-content/uploads/2014/05/Open_Licence.pdf}{French "Open Licence"}.
}
\seealso{
\link{french_departments} and \link{french_cities}
}
\keyword{datasets}
