% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_df.R
\name{sim_column}
\alias{sim_column}
\title{Get the collection of spatial interaction models from a sim data frame}
\usage{
sim_column(sim_df)
}
\arguments{
\item{sim_df}{a data frame of spatial interaction models, an object of class
\code{sim_df}}
}
\value{
the collection of spatial interaction models in the \code{sim_df} object,
as a \code{sim_list} object
}
\description{
Get the collection of spatial interaction models from a sim data frame
}
\examples{
positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
all_flows <- grid_blvim(distances, production, c(1.25, 1.5), c(1, 2, 3), attractiveness)
all_flows_df <- sim_df(all_flows)
sim_column(all_flows_df)
}
