% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_df.R
\name{sim_df_extract}
\alias{sim_df_extract}
\alias{$<-.sim_df}
\alias{[[<-.sim_df}
\alias{[.sim_df}
\alias{[<-.sim_df}
\alias{names<-.sim_df}
\title{Extract or replace parts of a SIM data frame}
\usage{
\method{$}{sim_df}(x, name) <- value

\method{[[}{sim_df}(x, i, j) <- value

\method{[}{sim_df}(x, i, j, ..., drop)

\method{[}{sim_df}(x, i, j) <- value

\method{names}{sim_df}(x) <- value
}
\arguments{
\item{x}{data frame of spatial interaction models, an object of class
\code{sim_df}}

\item{name}{a literal character string}

\item{value}{a suitable replacement value: it will be repeated a whole number
of times if necessary and it may be coerced: see the Coercion section in
\code{\link[base]{[.data.frame}} If NULL, deletes the column if a single
column is selected.}

\item{i, j, ...}{elements to extract or replace. For \code{[} and \code{[[}, these are
numeric or character or, for \code{[} only, empty or logical. Numeric values are
coerced to integer as if by as.integer. For replacement by \code{[}, a logical
matrix is allowed.}

\item{logical.}{If \code{TRUE} the result is coerced to the lowest possible
dimension. The default is to drop if only one column is left, but not to
drop if only one row is left.}
}
\value{
something
}
\description{
Extract or replace subsets of SIM data frames. The behaviour of the functions
is very close to the one of the corresponding \code{data.frame} functions, see
\code{\link[base]{[.data.frame}}. However, modifications of the SIM columns
or suppression of core columns will turn the object into a standard
\code{data.frame} to void issues in e.g. graphical representations, see below for
details.
}
\details{
In a \code{sim_df}, the core columns are derived from the \code{sim_list} column.
Replacement functions maintain this property by updating the columns after
any modification of the \code{sim_list} column. Modifications of the core columns
are rejected (removing a core column is accepted but this turns the \code{sim_df}
into a standard \code{data.frame}). Renaming the \code{sim_list} column is supported
but renaming a core column turns also the \code{sim_df}  into a standard
\code{data.frame}.

In addition, the \code{sim_list} column obeys to restriction on \code{sim_list}
modifications (i.e, a \code{sim_list} contains a homogeneous collection of spatial
interaction models).
}
