% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_fp_jacobian}
\alias{sim_fp_jacobian}
\title{Compute the Jacobian of the fixed-point map G}
\usage{
sim_fp_jacobian(sim, ...)
}
\arguments{
\item{sim}{a spatial interaction model, an object of class \code{sim}}

\item{...}{additional parameters}
}
\value{
a p x p matrix or NA
}
\description{
The function computes the Jacobian matrix of the fixed-point map with respect
to the attractivenesses. The function returns \code{NA} when applied to spatial
interaction models for which the map is not defined.
}
\details{
The fixed point map \eqn{\mathbf{G}} is given by

\deqn{G_j=D_j-\kappa_j Z_j,}

where \eqn{D_j} is the destination flow (\code{\link[=destination_flow]{destination_flow()}}),
\eqn{Z_j} the attractiveness (\code{\link[=attractiveness]{attractiveness()}}) and \eqn{\kappa_j} the
conversion factor (\code{\link[=sim_conversion]{sim_conversion()}}), all attached to destination \eqn{j}.

The derivatives of \eqn{\mathbf{G}} w.r.t. \eqn{\mathbf{Z}} can be used to
analyse the stability of the solutions, among other applications.
}
\examples{
distances <- french_cities_distances[1:10, 1:10] / 1000 ## convert to km
production <- log(french_cities$population[1:10])
attractiveness <- log(french_cities$area[1:10])
model <- static_blvim(distances, production, 1.5, 1 / 250, attractiveness)
sim_fp_jacobian(model) ## must NA
}
