% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_iterations}
\alias{sim_iterations}
\title{Returns the number of iterations used to produce this spatial interaction model}
\usage{
sim_iterations(sim, ...)
}
\arguments{
\item{sim}{a spatial interaction model object}

\item{...}{additional parameters}
}
\value{
a number of iterations that may be one if the spatial interaction model
has been obtained using a static model (see \code{\link[=static_blvim]{static_blvim()}}).
}
\description{
Returns the number of iterations used to produce this spatial interaction model
}
\examples{
positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
model <- static_blvim(distances, production, 1.5, 1, attractiveness)
sim_iterations(model) ## must be one
}
\seealso{
\code{\link[=sim_converged]{sim_converged()}}
}
