% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_potential}
\alias{sim_potential}
\title{Compute the potential of a spatial interaction model}
\usage{
sim_potential(sim, ...)
}
\arguments{
\item{sim}{a spatial interaction model, an object of class \code{sim}}

\item{...}{additional parameters}
}
\value{
the scalar value of the potential
}
\description{
The potential is a measure of stability introduced by Osawa, Akamatsu, and
Kogure. It is only defined for spatial interaction models produced by
BLV framework. The function returns \code{NA} when applied to spatial interaction
models for which a potential cannot be computed.
}
\details{
The potential of a spatial interaction model is given by

\deqn{\frac{1}{\alpha}\sum_{i=1}^n X_i \log\left(\sum_{j=1}^p Z_j^{\alpha}\,
\exp(-\beta c_{ij})\right)-\sum_{j=1}^p\kappa_jZ_j,}

where:
\itemize{
\item \eqn{c} is the cost matrix \code{\link[=costs]{costs()}}
\item \eqn{X} is the production constraint vector (\code{\link[=production]{production()}})
\item \eqn{\alpha} is the return to scale parameter (\code{\link[=return_to_scale]{return_to_scale()}})
\item \eqn{\beta} is the inverse of a cost scale parameter (\code{\link[=inverse_cost]{inverse_cost()}})
\item \eqn{Z} is the attractiveness vector (\code{\link[=attractiveness]{attractiveness()}})
\item \eqn{\kappa} is the conversion factor vector (\code{\link[=sim_conversion]{sim_conversion()}})
}
}
\examples{
distances <- french_cities_distances[1:10, 1:10] / 1000 ## convert to km
production <- log(french_cities$population[1:10])
attractiveness <- log(french_cities$area[1:10])
model <- static_blvim(distances, production, 1.5, 1 / 250, attractiveness)
sim_conversion(model) ## must NA
}
\references{
Osawa, M., Akamatsu, T., & Kogure, Y. (2025). "Most likely retail
agglomeration patterns: Potential maximization and stochastic stability of
spatial equilibria". \url{https://arxiv.org/abs/2011.06778v2}
}
