% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_list_summary.R
\name{summary.sim_list}
\alias{summary.sim_list}
\title{Summary of a collection of spatial interaction models}
\usage{
\method{summary}{sim_list}(object, ...)
}
\arguments{
\item{object}{a collection of spatial interaction models, an object of class
\code{sim_list}}

\item{...}{additional parameters (not used currently)}
}
\value{
a list with a set of summary statistics computed on the collection
of spatial interaction models
}
\description{
Summary of a collection of spatial interaction models
}
\examples{
positions <- matrix(rnorm(15 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 15)
attractiveness <- rep(1, 15)
all_flows <- grid_blvim(distances,
  production,
  c(1.1, 1.25, 1.5),
  c(1, 2, 3),
  attractiveness,
  epsilon = 0.1,
  bipartite = FALSE,
)
summary(all_flows)
}
