% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_autoplot.R
\name{autoplot.sim}
\alias{autoplot.sim}
\title{Create a complete ggplot for a spatial interaction model}
\usage{
\method{autoplot}{sim}(object, flows = c("full", "destination", "attractiveness"), ...)
}
\arguments{
\item{object}{a spatial interaction model object}

\item{flows}{\code{"full"} (default),  \code{"destination"} or \code{"attractiveness"}, see details.}

\item{...}{additional parameters (not used currently)}
}
\value{
a ggplot object
}
\description{
This function represents graphical the flows of a spatial interaction model,
in different direct or aggregated forms.
}
\details{
The graphical representation depends on the value of \code{flows}:
\itemize{
\item \code{"full"}: this is the default case for which the full flow matrix is represented.
It is extracted from the spatial interaction model with \code{\link[=flows]{flows()}} and
displayed using a matrix representation with origin locations in rows and
destination locations in columns. The colour of a cell corresponds to the
intensity of a flow between the corresponding locations. To mimic the
standard top to bottom reading order of a flow matrix, the top row of the
graphical representation corresponds to the first origin location.
\item \code{"destination"}: the function computes the
incoming flows for destination locations (using \code{\link[=destination_flow]{destination_flow()}}) and
represents them with a bar plot (each bar is proportional to the incoming
flow);
\item \code{"attractiveness"}: the function uses a bar plot to represent the
attractivenesses of the destination locations (as given by \code{\link[=attractiveness]{attractiveness()}}).
This is interesting for dynamic models where those values are updated during
the iterations (see \code{\link[=blvim]{blvim()}} for details). When the calculation has
converged (see \code{\link[=sim_converged]{sim_converged()}}), both \code{"destination"} and \code{"attractiveness"} graphics should
be almost identical.
}
}
\examples{
positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
flows <- blvim(distances, production, 1.5, 4, attractiveness)
ggplot2::autoplot(flows)
## bar plots should be almost identical if convergence occured
sim_converged(flows)
ggplot2::autoplot(flows, "destination")
ggplot2::autoplot(flows, "attractiveness")
}
