% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_list_autoplot.R
\name{autoplot.sim_list}
\alias{autoplot.sim_list}
\title{Create a complete ggplot for a collection of spatial interaction models}
\usage{
\method{autoplot}{sim_list}(
  object,
  statistics = c("shannon", "renyi", "terminals", "iterations"),
  inverse = TRUE,
  order = 1,
  definition = c("ND", "RW"),
  ...
)
}
\arguments{
\item{object}{a collection of spatial interaction models, an object of class
\code{sim_list}}

\item{statistics}{the value to display, defaulting to \code{"shannon"}, see
details}

\item{inverse}{whether to use the cost scale parameter (default)}

\item{order}{order of the Rényi entropy, used only when \code{definition="renyi"}}

\item{definition}{diversity definition \code{"shannon"} (default),
\code{"renyi"} (see details) or a definition supported by  \code{\link[=terminals]{terminals()}}}

\item{...}{additional parameters (not used currently)}
}
\value{
a ggplot object
}
\description{
This function uses a tile plot from ggplot2 to display a single numerical
value for each of the parameter pairs used to produce the collection of
spatial interaction models.
}
\details{
The value to display is selected with the \code{statistics} parameter and defaults
to the Shannon diversity with \code{"shannon"} (see \code{\link[=diversity]{diversity()}}). Othervalues
include:
\itemize{
\item \code{"renyi"} (coupled with the \code{order} parameter) for Rényi diversity (see
\code{\link[=diversity]{diversity()}})
\item \code{"terminals"} (coupled with the \code{definition} parameter) for the number of
terminals when the origin and destination locations are identical (see
\code{\link[=terminals]{terminals()}})
\item \code{"iterations"} for the number of iterations the \verb{[blvim()]} model took
to converge
}

The horizontal axis is used by default for the cost scale parameter, that is
\eqn{1/\beta}. This is in general easier to read than using the inverse cost
scale. The \code{inverse} parameter can be used to turn off this feature. The
vertical axis is used by default for the return to scale parameter.
}
\examples{
positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
all_flows <- grid_blvim(
  distances,
  production,
  c(1.25, 1.5, 1.75),
  1 / c(0.25, 0.5, 1),
  attractiveness
)
ggplot2::autoplot(all_flows)
}
\seealso{
\code{\link[=grid_blvim]{grid_blvim()}}, \code{\link[=diversity]{diversity()}}, \code{\link[=terminals]{terminals()}}
}
