% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{blvim}
\alias{blvim}
\title{Compute an equilibrium solution of the Boltzmann–Lotka–Volterra model}
\usage{
blvim(
  costs,
  X,
  alpha,
  beta,
  Z,
  epsilon = 0.01,
  iter_max = 50000,
  conv_check = 100,
  precision = 1e-06,
  quadratic = FALSE
)
}
\arguments{
\item{costs}{a cost matrix}

\item{X}{a vector of production constraints}

\item{alpha}{the return to scale parameter}

\item{beta}{the inverse cost scale parameter}

\item{Z}{a vector of initial destination attractivenesses}

\item{epsilon}{the update intensity}

\item{iter_max}{the maximal number of steps of the BLV dynamic}

\item{conv_check}{number of iterations between to convergence test}

\item{precision}{convergence threshold}

\item{quadratic}{selects the update rule, see details.}
}
\value{
an object of class \code{sim}(and \code{sim_blvim}) for spatial interaction
model that contains the matrix of flows between the origin and the
destination locations as well as the final attractivenesses computed by the
model.
}
\description{
This function computes flows between origin locations and destination
locations at an equilibrium solution of A. Wilson's Boltzmann–Lotka–Volterra
(BLV) interaction model. The BLV dynamic model is initialised with the
production constraints at the origin locations and the initial values of the
the attractiveness of destination locations. Iterations update the
attractivenesses according the received flows.
}
\details{
In a step of the BLV model, flows are computed according to the production
constrained entropy maximising model proposed by A. Wilson and implemented in
\code{\link[=static_blvim]{static_blvim()}}. Then the flows received at a destination are computed as
follows

\deqn{\forall j,\quad D_j=\sum_{i=1}^{n}Y_{ij},}

for destination \eqn{j}. This enables updating the attractivenesses by making
them closer to the received flows, i.e. trying to reduce \eqn{|D_j-Z_j|}.

A. Wilson and co-authors proposed two different update strategies:
\enumerate{
\item The original model proposed in Harris & Wilson (1978) updates the
\eqn{Z_j} as follows \deqn{Z_j^{t+1} = Z_j^{t} + \epsilon (D^{t}_j-Z^{t}_j)}
\item In Wilson (2008), the update is given by \deqn{Z_j^{t+1} = Z_j^{t} +
\epsilon (D^{t}_j-Z^{t}_j)Z^{t}_j}
}

In both cases, \eqn{\epsilon} is given by the \code{epsilon} parameter. It should
be smaller than 1. The first update is used when the \code{quadratic} parameter is
\code{FALSE} which is the default value. The second update is used when
\code{quadratic} is \code{TRUE}.

Updates are performed until convergence or for a maximum of \code{iter_max}
iterations. Convergence is checked every \code{conv_check} iterations. The
algorithm is considered to have converged if \deqn{\|Z^{t+1}-Z^t\|<\delta
(\|Z^{t+1}\|+\delta),} where \eqn{\delta} is given by the \code{precision}
parameter.
}
\examples{
positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
flows <- blvim(distances, production, 1.5, 1, attractiveness)

}
\references{
Harris, B., & Wilson, A. G. (1978). "Equilibrium Values and
Dynamics of Attractiveness Terms in Production-Constrained
Spatial-Interaction Models", Environment and Planning A: Economy and Space,
10(4), 371-388. \doi{10.1068/a100371}

Wilson, A. (2008), "Boltzmann, Lotka and Volterra and spatial structural
evolution: an integrated methodology for some dynamical systems", J. R.
Soc. Interface.5865–871 \doi{10.1098/rsif.2007.1288}
}
