% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_list_common.R
\name{costs.sim_list}
\alias{costs.sim_list}
\title{Extract the common cost matrix from a collection of spatial interaction models}
\usage{
\method{costs}{sim_list}(sim, ...)
}
\arguments{
\item{sim}{a collection of spatial interaction models, an object of class
\code{sim_list}}

\item{...}{additional parameters}
}
\value{
the cost matrix
}
\description{
Extract the common cost matrix from a collection of spatial interaction models
}
\examples{
positions <- matrix(rnorm(15 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 15)
attractiveness <- rep(1, 15)
all_flows <- grid_blvim(distances,
  production,
  c(1.1, 1.25, 1.5),
  c(1, 2, 3),
  attractiveness,
  epsilon = 0.1
)
## should be TRUE
identical(distances, costs(all_flows))

}
\seealso{
\code{\link[=costs]{costs()}}, \code{\link[=grid_blvim]{grid_blvim()}}
}
