% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{destination_flow}
\alias{destination_flow}
\title{Compute the flows incoming at each destination location}
\usage{
destination_flow(sim, ...)
}
\arguments{
\item{sim}{a spatial interaction model object}

\item{...}{additional parameters}
}
\value{
a vector of flows incoming at destination locations
}
\description{
Compute the flows incoming at each destination location
}
\examples{
positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
model <- static_blvim(distances, production, 1.5, 1, attractiveness)
destination_flow(model)
}
