% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_converged}
\alias{sim_converged}
\title{Reports whether the spatial interaction model construction converged}
\usage{
sim_converged(sim, ...)
}
\arguments{
\item{sim}{a spatial interaction model object}

\item{...}{additional parameters}
}
\value{
\code{TRUE}, \code{FALSE} or \code{NA}, as described above
}
\description{
Some spatial interaction models are the result of an iterative calculation,
see for instance \code{\link[=blvim]{blvim()}}. This calculation may have been interrupted before
convergence. The present function returns \code{TRUE} if the calculation converged,
\code{FALSE} if this was not the case and \code{NA} if the spatial interaction model
is not the result of an interative calculation .
}
\examples{
positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
model <- static_blvim(distances, production, 1.5, 1, attractiveness)
sim_converged(model) ## must be NA
}
\seealso{
\code{\link[=sim_iterations]{sim_iterations()}}
}
