% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_conversion}
\alias{sim_conversion}
\title{Returns the conversion factors between attractivenesses and incoming flows}
\usage{
sim_conversion(sim, ...)
}
\arguments{
\item{sim}{a spatial interaction model object (an object of class \code{sim})}

\item{...}{additional parameters}
}
\value{
NA or a vector of conversion factors, see above.
}
\description{
Some spatial interaction models use conversion factors between attractivenesses
and incoming flows to check whether they are at equilibrium (see for instance
\code{\link[=blvim]{blvim()}}). This function returns the conversion factors used to build this
sim is applicable, and \code{NA} when not applicable.
}
\examples{
distances <- french_cities_distances[1:10, 1:10] / 1000 ## convert to km
production <- log(french_cities$population[1:10])
attractiveness <- log(french_cities$area[1:10])
model <- static_blvim(distances, production, 1.5, 1 / 250, attractiveness)
sim_conversion(model) ## must NA
}
\seealso{
\code{\link[=blvim]{blvim()}}
}
