% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_df.R
\name{sim_df}
\alias{sim_df}
\title{Create a spatial interaction data frame from a collection of interaction models}
\usage{
sim_df(x)
}
\arguments{
\item{x}{a collection of spatial interaction models, an object of class
\code{sim_list}}
}
\value{
a data frame representation of the spatial interaction model collection
with classes \code{sim_df} and \code{data.frame}
}
\description{
This function build a data frame from a collection of spatial interaction
model. The data frame has a list column \code{sim} which stores the collection
of models and classical columns that contain characteristics of the models.
See details for the default columns.
}
\details{
The data frame has one row per spatial interaction model and the following columns:
\itemize{
\item \code{sim}: the last column that contains the model
\item \code{alpha}: the return to scale parameter used to build the model
\item \code{beta}: the cost inverse scale parameter used to build the model
\item \code{diversity}: model default \code{\link[=diversity]{diversity()}} (Shannon's diversity)
\item \code{iterations}: the number of iterations used to produce the model (1 for a
static model)
\item \code{converged}: \code{TRUE} is the iterative calculation of the model converged (for
models produced by \code{\link[=blvim]{blvim()}} and related approaches), \code{FALSE} for no convergence
and \code{NA} for static models
}
}
\examples{
positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
all_flows <- grid_blvim(distances, production, c(1.25, 1.5), c(1, 2, 3), attractiveness)
all_flows_df <- sim_df(all_flows)
all_flows_df$converged

}
