% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{static_blvim}
\alias{static_blvim}
\title{Compute flows between origin and destination locations}
\usage{
static_blvim(costs, X, alpha, beta, Z)
}
\arguments{
\item{costs}{a cost matrix}

\item{X}{a vector of production constraints}

\item{alpha}{the return to scale parameter}

\item{beta}{the cost inverse scale parameter}

\item{Z}{a vector of destination attractivenesses}
}
\value{
a matrix of flows between the origin and the destination locations
(see \eqn{(Y_{ij})_{1\leq i\leq n, 1\leq j\leq p}} above).
}
\description{
This function computes flows between origin locations and destination
locations according to the production constrained entropy maximising model
proposed by A. Wilson.
}
\details{
The model computes flows using the following parameters:
\itemize{
\item \code{costs} (\eqn{c}) is a \eqn{n\times p} matrix whose \eqn{(i,j )} entry is the
cost of a "unit" flow from origin location \eqn{i} to destination location
\eqn{j}
\item \code{X} (\eqn{X}) is a vector of size \eqn{n} containing non negative production
constraints for the \eqn{n} origin locations
\item \code{alpha} (\eqn{\alpha}) is a return to scale parameter that enhance (or reduce is smaller
that 1) the attractiveness of destination location larger than 1
\item \code{beta} (\eqn{\beta}) is the inverse of a cost scale parameter, i.e., costs are multiplied
by \code{beta} in the model
\item \code{Z} (\eqn{Z}) is a vector of size \eqn{p} containing positive attractivenesses for
the \eqn{p} destination locations
}

According to Wilson's model, the flow between origin location \eqn{i} and
destination location \eqn{j}, \eqn{Y_{ij}}, is given by

\deqn{Y_{ij}=\frac{X_iZ_j^{\alpha}\exp(-\beta
c_{ij})}{\sum_{k=1}^pZ_k^{\alpha}\exp(-\beta c_{ik})}.}

The model is production constrained because

\deqn{\forall i,\quad X_i=\sum_{j=1}^{p}Y_{ij},}

that is the origin location \eqn{i} sends a total flow of exactly \eqn{X_i}.
}
\examples{
positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
flows <- static_blvim(distances, production, 1.5, 1, attractiveness)

}
\references{
Wilson, A. (1971), "A family of spatial interaction models, and
associated developments", Environment and Planning A: Economy and Space,
3(1), 1-32 \doi{10.1068/a030001}
}
